requirements.txt
numpy==1.11.2
pandas==0.19.0
Pillow==3.4.1
protobuf==3.0.0
psycopg2==2.6.2
python-dateutil==2.5.3
pytz==2016.7
six==1.10.0
tensorflow==0.11.0rc2
